Refactor @stacks imports into stacks module
