--- src/assets/constants.js
@@ -1,7 +1,0 @@
-import { AppConfig } from '@stacks/auth';
-
-export const appConfig = new AppConfig(['store_write', 'publish_data']);
-
-export const TASKS_FILENAME = 'tasks.json';
-
-export const EXPLORER_URL = 'https://explorer.blockstack.org';
